refactor(cli): split login and logout with shared context
